cleanup: break utils files into output/input/prompts files for commands
